+++ packages/interactions/route-prefetch/types/index.d.ts
@@ -1,6 +1,6 @@
-import { Interaction, MatchResponseProperties } from "@curi/types";
+import { Interaction, Match } from "@curi/types";
 export interface PrefetchCallOptions {
-    match?: MatchResponseProperties;
+    match?: Match;
     external?: any;
 }
 export default function prefetch_route(): Interaction;
