+++ packages/react-dom/.size-snapshot.json
@@ -19,13 +19,13 @@
     "gzipped": 953
   },
   "dist/curi-react-dom.umd.js": {
-    "bundled": 10804,
-    "minified": 4071,
-    "gzipped": 1726
+    "bundled": 10817,
+    "minified": 4084,
+    "gzipped": 1730
   },
   "dist/curi-react-dom.min.js": {
-    "bundled": 10774,
-    "minified": 4041,
-    "gzipped": 1709
+    "bundled": 10787,
+    "minified": 4054,
+    "gzipped": 1713
   }
 }
