+++ packages/router/src/index.ts
@@ -1,5 +1,3 @@
-export * from "./types";
-
 import create_router from "./create_router";
 import prepare_routes from "./prepare_routes";
