--- packages/router/src/types/index.ts
@@ -1,10 +1,0 @@
-export {
-  Route,
-  SyncRoute,
-  AsyncRoute,
-  RouteDescriptor,
-  AsyncMatchFn,
-  ResolveResults
-} from "./route";
-export { RawParams } from "./response";
-export { RouterOptions } from "./curi";
