--- packages/router/src/types/response.ts
@@ -1,1 +1,0 @@
-export type RawParams = { [key: string]: string };
