+++ packages/router/types/create_router.d.ts
@@ -1,4 +1,3 @@
 import { HistoryConstructor, HistoryOptions } from "@hickory/root";
-import { PreparedRoutes, CuriRouter } from "@curi/types";
-import { RouterOptions } from "./types/curi";
+import { PreparedRoutes, CuriRouter, RouterOptions } from "@curi/types";
 export default function create_router<HOpts = HistoryOptions>(history_constructor: HistoryConstructor<HOpts>, route_array: PreparedRoutes, options?: RouterOptions<HOpts>): CuriRouter;
