+++ packages/router/types/index.d.ts
@@ -1,4 +1,3 @@
-export * from "./types";
 import create_router from "./create_router";
 import prepare_routes from "./prepare_routes";
 export { create_router, prepare_routes };
