+++ packages/router/types/match_location.d.ts
@@ -1,4 +1,13 @@
 import { SessionLocation } from "@hickory/root";
-import { PreparedRoute } from "@curi/types";
-import { PossibleMatch } from "./types/match";
+import { PreparedRoute, Match } from "@curi/types";
+interface MissMatch {
+    route: undefined;
+    match: undefined;
+}
+export interface RealMatch {
+    route: PreparedRoute;
+    match: Match;
+}
+export declare type PossibleMatch = RealMatch | MissMatch;
 export default function match_location(location: SessionLocation, routes: Array<PreparedRoute>): PossibleMatch;
+export {};
