+++ packages/router/types/prepare_routes.d.ts
@@ -1,3 +1,2 @@
-import { PreparedRoutes } from "@curi/types";
-import { RouteDescriptor } from "./types/route";
+import { PreparedRoutes, RouteDescriptor } from "@curi/types";
 export default function prepare_routes(user_routes: Array<RouteDescriptor>): PreparedRoutes;
