+++ packages/router/types/resolve_matched_route.d.ts
@@ -1,3 +1,3 @@
-import { Match } from "./types/match";
-import { ResolveResults } from "./types/route";
-export default function resolve_route(match: Match, global: any): Promise<ResolveResults>;
+import { ResolveResults } from "@curi/types";
+import { RealMatch } from "./match_location";
+export default function resolve_route(match: RealMatch, global: any): Promise<ResolveResults>;
