--- packages/router/types/types/index.d.ts
@@ -1,3 +1,0 @@
-export { Route, SyncRoute, AsyncRoute, RouteDescriptor, AsyncMatchFn, ResolveResults } from "./route";
-export { RawParams } from "./response";
-export { RouterOptions } from "./curi";
