--- packages/router/types/types/response.d.ts
@@ -1,3 +1,0 @@
-export declare type RawParams = {
-    [key: string]: string;
-};
