--- packages/router/types/utils/parse_params.d.ts
@@ -1,3 +1,0 @@
-import { ParamParsers, Params } from "@curi/types";
-import { RawParams } from "../types/response";
-export default function parse_params(params: RawParams, fns?: ParamParsers): Params;
