+++ packages/types/src/index.ts
@@ -1,16 +1,41 @@
-import { PathFunction, Key } from "path-to-regexp";
+import {
+  PathFunction,
+  PathFunctionOptions,
+  RegExpOptions,
+  Key
+} from "path-to-regexp";
 import {
   History,
+  HistoryOptions,
   SessionLocation,
   PartialLocation,
   Action,
   NavType
 } from "@hickory/root";
 
-// router
+// a route descriptor comes from the user
+export interface RouteDescriptor {
+  name: string;
+  path: string;
+  path_options?: RegExpOptions;
+  params?: ParamParsers;
+  children?: Array<RouteDescriptor>;
+  response?: ResponseFn;
+  resolve?: AsyncMatchFn;
+  extra?: { [key: string]: any };
+}
 
-export type Observer = (props?: Emitted) => void;
+// third argument to create_router
+export interface RouterOptions<O = HistoryOptions> {
+  route?: Array<Interaction>;
+  side_effects?: Array<Observer>;
+  pathname_options?: PathFunctionOptions;
+  emit_redirects?: boolean;
+  external?: any;
+  history?: O;
+}
 
+// object returned by create_router
 export interface CuriRouter {
   refresh: (routes?: PreparedRoutes) => void;
   observe: (fn: Observer, options?: ResponseHandlerOptions) => RemoveObserver;
@@ -23,30 +48,47 @@ export interface CuriRouter {
   external: any;
 }
 
+// options passed to router.navigate
 export interface NavigationDetails extends RouteLocation {
   method?: NavType;
   cancelled?: () => void;
   finished?: () => void;
 }
-
+// callback function to call when cancelling an async navigation
 export type CancelNavigateCallbacks = () => void;
 
+// the current response and navigation
 export interface CurrentResponse {
   response: Response | null;
   navigation: Navigation | null;
 }
+// information about a navigation
+export interface Navigation {
+  action: Action;
+  previous: Response | null;
+}
 
+// configuration options for router.observe and router.once
 export interface ResponseHandlerOptions {
   initial?: boolean;
 }
+// an observer function that will be called when there is a new navigation
+export type Observer = (props?: Emitted) => void;
+// observers registered with router.observe can be removed
 export type RemoveObserver = () => void;
 
-export type CancelActiveNavigation = () => void;
+// a function to
 export type Cancellable = (cancel?: CancelActiveNavigation) => void;
+export type CancelActiveNavigation = () => void;
 export type RemoveCancellable = () => void;
 
-// location object
+export interface Emitted {
+  response: Response;
+  navigation: Navigation;
+  router: CuriRouter;
+}
 
+// a description of a route to navigate to
 export interface RouteLocation {
   name?: string;
   params?: Params;
@@ -55,24 +97,11 @@ export interface RouteLocation {
   state?: any;
 }
 
-// emit
-
-export interface Navigation {
-  action: Action;
-  previous: Response | null;
-}
-
-export interface Emitted {
-  response: Response;
-  navigation: Navigation;
-  router: CuriRouter;
-}
-
 // response
 
 export type Params = { [key: string]: any };
 
-export interface MatchResponseProperties {
+export interface Match {
   location: SessionLocation;
   name: string;
   params: Params;
@@ -85,7 +114,7 @@ export interface RedirectLocation extends PartialLocation {
   url: string;
 }
 
-export interface Response extends MatchResponseProperties {
+export interface Response extends Match {
   status?: number;
   error?: any;
   body?: any;
@@ -106,16 +135,15 @@ export interface Route<R = unknown> {
   pathname: PathFunction;
   resolve: R;
 }
+export interface SyncRoute extends Route<undefined> {}
+export interface AsyncRoute extends Route<AsyncMatchFn> {}
 
-export type ResponseFn = (
-  props: Readonly<ResponseBuilder>
-) => SettableResponseProperties;
+export type AsyncMatchFn = (
+  matched?: Readonly<Match>,
+  external?: any
+) => Promise<any>;
 
-export interface PathMatching {
-  exact: boolean;
-  re: RegExp;
-  keys: Array<Key>;
-}
+// prepared routes
 
 export interface PreparedRoute {
   public: Route;
@@ -125,6 +153,16 @@ export interface PreparedRoute {
   path_matching: PathMatching;
   param_parsers?: ParamParsers;
 }
+
+export interface PathMatching {
+  exact: boolean;
+  re: RegExp;
+  keys: Array<Key>;
+}
+
+export type ResponseFn = (
+  props: Readonly<ResponseBuilder>
+) => SettableResponseProperties;
 
 export type PreparedRoutes = Array<PreparedRoute>;
 
@@ -149,8 +187,14 @@ export interface ParamParsers {
 export interface ResponseBuilder {
   resolved: any;
   error: any;
-  match: MatchResponseProperties;
+  match: Match;
   external: any;
+}
+
+// should this exist?
+export interface ResolveResults {
+  resolved: any;
+  error: any;
 }
 
 // interaction
